Parameter detection via regex has been removed
The problem with detecting the payload requirements of a bound function
by using its parameter names as repository lookup keys is that they will
get clobbered when your code is compressed. To address this, the whole
library has been rewritten to use defined parameters in a passed array
when binding a method.